Extract messaging logic into separate package. Create customers service related messages
