+++ internal/communication/customers/customers.go
@@ -1,0 +1,29 @@
+package customers
+
+import "github.com/fdemchenko/exchanger/internal/communication"
+
+const QueueName = "emails"
+
+const (
+	CreateCustomerRequestQueue  = "CreateCustomerRequests"
+	CreateCustomerResponseQueue = "CreateCustomerResponces"
+)
+
+const (
+	CreateCustomerRequest  communication.MessageType = "CreateCustomerRequest"
+	CustomerCreated        communication.MessageType = "CustomerCreated"
+	CustomerCreationFailed communication.MessageType = "CustomerCreationFailed"
+)
+
+type CreateCustomerRequestPayload struct {
+	Email          string `json:"email"`
+	SubscriptionID int    `json:"int"`
+}
+
+type CustomerCreatedPayload struct {
+	ID int `json:"id"`
+}
+
+type CustomerCreationFailedPayload struct {
+	Error string `json:"error"`
+}
