--- internal/communication/mailer/queue.go
@@ -1,3 +1,0 @@
-package mailer
-
-const QueueName = "emails"
