+++ internal/communication/message.go
@@ -1,0 +1,17 @@
+package communication
+
+import (
+	"time"
+)
+
+type MessageType string
+
+type MessageHeader struct {
+	Type      MessageType `json:"messageType"`
+	Timestamp time.Time   `json:"timestamp"`
+}
+
+type Message[T any] struct {
+	MessageHeader
+	Payload T `json:"payload"`
+}
