+++ internal/services/scheduler.go
@@ -5,6 +5,7 @@ import (
 	"encoding/json"
 	"time"
 
+	"github.com/fdemchenko/exchanger/internal/communication"
 	"github.com/fdemchenko/exchanger/internal/communication/mailer"
 	amqp "github.com/rabbitmq/amqp091-go"
 	"github.com/rs/zerolog/log"
@@ -69,8 +70,8 @@ func (es *RabbitMQEmailScheduler) sendMessages() error {
 	if err != nil {
 		return err
 	}
-	rateUpdateMessage := mailer.Message[mailer.ExchangeRateUpdatedEvent]{
-		MessageHeader: mailer.MessageHeader{Type: mailer.ExchangeRateUpdated, Timestamp: time.Now()},
+	rateUpdateMessage := communication.Message[mailer.ExchangeRateUpdatedEvent]{
+		MessageHeader: communication.MessageHeader{Type: mailer.ExchangeRateUpdated, Timestamp: time.Now()},
 		Payload:       mailer.ExchangeRateUpdatedEvent{Rate: rate},
 	}
 	bytes, err := json.Marshal(rateUpdateMessage)
@@ -93,8 +94,8 @@ func (es *RabbitMQEmailScheduler) sendMessages() error {
 	}
 
 	for _, email := range emails {
-		sendEmailMessage := mailer.Message[mailer.SendEmailNotificationCommand]{
-			MessageHeader: mailer.MessageHeader{Type: mailer.SendEmailNotification, Timestamp: time.Now()},
+		sendEmailMessage := communication.Message[mailer.SendEmailNotificationCommand]{
+			MessageHeader: communication.MessageHeader{Type: mailer.SendEmailNotification, Timestamp: time.Now()},
 			Payload:       mailer.SendEmailNotificationCommand{Email: email},
 		}
 		bytes, err := json.Marshal(sendEmailMessage)
